ai_microservice/requirements.txt
fastapi
uvicorn
face_recognition
numpy
pillow
easyocr
pydantic
opencv-python
python-multipart
